// src/types/competitor.ts
export interface CompetitorData {
  keyword: string;
  ad_area: 'PC' | 'Mobile';
  advertiser: string;
  url: string;
  average: number;
  hour_00: number;
  hour_01: number;
  hour_02: number;
  hour_03: number;
  hour_04: number;
  hour_05: number;
  hour_06: number;
  hour_07: number;
  hour_08: number;
  hour_09: number;
  hour_10: number;
  hour_11: number;
  hour_12: number;
  hour_13: number;
  hour_14: number;
  hour_15: number;
  hour_16: number;
  hour_17: number;
  hour_18: number;
  hour_19: number;
  hour_20: number;
  hour_21: number;
  hour_22: number;
  hour_23: number;
}

export interface AnalysisResult {
  keyword: string;
  ad_area: 'PC' | 'Mobile';
  advertiser: string;
  url: string;
  average: number;
  peak_hour: string;
  peak_value: number;
  lowest_hour: string;
  lowest_value: number;
  active_hours: number;
  hour_00: number;
  hour_01: number;
  hour_02: number;
  hour_03: number;
  hour_04: number;
  hour_05: number;
  hour_06: number;
  hour_07: number;
  hour_08: number;
  hour_09: number;
  hour_10: number;
  hour_11: number;
  hour_12: number;
  hour_13: number;
  hour_14: number;
  hour_15: number;
  hour_16: number;
  hour_17: number;
  hour_18: number;
  hour_19: number;
  hour_20: number;
  hour_21: number;
  hour_22: number;
  hour_23: number;
}